my eco> 내 댓글 읽기 > 댓글 게시글 조회

+++ eco100-app/app/src/main/java/com/mapo/eco100/entity/myeco/MyComment.kt
@@ -3,8 +3,8 @@ package com.mapo.eco100.entity.myeco
 import java.io.Serializable
 
 data class MyComment(
-    val commentId: Int,
-    val boardId: Int,
+    val commentId: Long,
+    val boardId: Long,
     val writer: String,
     val date: String,
     val contents: String
